LocalRegistrator

Documentation for LocalRegistrator.

LocalRegistrator.head_commit_hash

LocalRegistrator.isdirty

LocalRegistrator.istoplevel

LocalRegistrator.register

LocalRegistrator.tree_hash

LocalRegistrator.head_commit_hash — Method

head_commit_hash(path::String)

Returns commit hash of HEAD in repository located at path. The repository must exist and be a top-level repository, see istoplevel

LocalRegistrator.isdirty — Method

isdirty(path::String)

Returns true if the repository located at path is dirty and false otherwise. It does not throw an exception if the git process fails.

LocalRegistrator.istoplevel — Method

istoplevel(path::String)

Returns true if a top-level repository is located at path and false otherwise. It does not throw an exception if the git process fails.

LocalRegistrator.register — Method

register(; project_url::S, registry_url::S, registry_deps::Vector{S}=["https://github.com/JuliaRegistries/General.git"], dry_run::Bool=false) where {S<:AbstractString}

Register lastest version of a package located at project_url (in best case a remote repository) in a local registry located at registry_url, e.g.

register(; project_url="https://github.com/SuiteSplines/SuiteSplinesExamplePkg.jl.git", registry_url="https://github.com/SuiteSplines/SuiteSplinesRegistry.git")

If your package depends on some other local repositories, include these in registry_deps. The General repository is included by default.

LocalRegistrator.tree_hash — Method

tree_hash(path::String; commit::String = "HEAD")

Returns tree hash of commit in repository located at path. The repository must exist and be a top-level repository, see istoplevel